requirements.txt
anyio==4.9.0
arrow==1.3.0
async-timeout==5.0.1
attrs==25.3.0
backoff==2.2.1
backports-datetime-fromisoformat==2.0.3
blinker==1.9.0
cachelib==0.13.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
Deprecated==1.2.18
ecdsa==0.19.1
exceptiongroup==1.3.0
Flask==3.1.1
Flask-Caching==2.3.1
Flask-Limiter==3.11.0
flask-marshmallow==1.3.0
Flask-SQLAlchemy==3.1.1
flask-swagger==0.2.14
flask-swagger-ui==5.21.0
fqdn==1.5.1
graphql-core==3.2.6
h11==0.16.0
harfile==0.3.0
httpcore==1.0.9
httpx==0.28.1
hypothesis==6.135.29
hypothesis-graphql==0.11.1
hypothesis-jsonschema==0.23.1
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
isoduration==20.11.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
junit-xml==1.9
limits==4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
marshmallow-sqlalchemy==1.4.2
mdurl==0.1.2
mysql-connector-python==9.3.0
ordered-set==4.1.0
packaging==24.2
pluggy==1.6.0
pyasn1==0.6.1
Pygments==2.19.1
pyrate-limiter==3.7.1
pytest==8.4.1
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-subtests==0.14.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
requests==2.32.4
rfc3339-validator==0.1.4
rfc3987==1.3.8
rich==13.9.4
rpds-py==0.26.0
rsa==4.9.1
schemathesis==4.0.9
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.41
starlette==0.47.1
starlette-testclient==0.4.1
tomli==2.2.1
types-python-dateutil==2.9.0.20250708
typing_extensions==4.13.2
uri-template==1.3.0
urllib3==2.5.0
webcolors==24.11.1
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.22.0
